fix: duplicate trpc calls
