+++ apps/dashboard/src/app/(dashboard)/monitors/[id]/layout.tsx
@@ -25,10 +25,10 @@ export default async function Layout({
   const { id } = await params;
   const queryClient = getQueryClient();
 
-  await fetchQueryOrNotFound(
-    trpc.monitor.get.queryOptions({ id: Number.parseInt(id) }),
-  );
   await Promise.all([
+    fetchQueryOrNotFound(
+      trpc.monitor.get.queryOptions({ id: Number.parseInt(id) }),
+    ),
     queryClient.prefetchQuery(trpc.notification.list.queryOptions()),
     queryClient.prefetchQuery(trpc.privateLocation.list.queryOptions()),
   ]);
