+++ apps/dashboard/src/app/(dashboard)/notifications/layout.tsx
@@ -4,7 +4,7 @@ import {
   AppHeaderContent,
 } from "@/components/nav/app-header";
 import { AppSidebarTrigger } from "@/components/nav/app-sidebar";
-import { HydrateClient, getQueryClient, trpc } from "@/lib/trpc/server";
+import { HydrateClient } from "@/lib/trpc/server";
 
 import { Breadcrumb } from "./breadcrumb";
 import { NavActions } from "./nav-actions";
@@ -14,8 +14,6 @@ export default async function Layout({
 }: {
   children: React.ReactNode;
 }) {
-  const queryClient = getQueryClient();
-  await queryClient.prefetchQuery(trpc.notification.list.queryOptions());
   return (
     <HydrateClient>
       <AppHeader>
