+++ apps/dashboard/src/app/(dashboard)/overview/layout.tsx
@@ -16,10 +16,7 @@ export default async function Layout({
 }) {
   const queryClient = getQueryClient();
 
-  // inputs must match page.tsx's queries exactly — a differing input misses the cache
   await Promise.all([
-    queryClient.prefetchQuery(trpc.monitor.list.queryOptions()),
-    queryClient.prefetchQuery(trpc.page.list.queryOptions()),
     queryClient.prefetchQuery(trpc.incident.list.queryOptions()),
     queryClient.prefetchQuery(trpc.statusReport.list.queryOptions({})),
     queryClient.prefetchQuery(trpc.maintenance.list.queryOptions()),
