+++ apps/dashboard/src/app/(dashboard)/settings/sso/layout.tsx
@@ -4,20 +4,13 @@ import {
   AppHeaderContent,
 } from "@/components/nav/app-header";
 import { AppSidebarTrigger } from "@/components/nav/app-sidebar";
-import { HydrateClient, getQueryClient, trpc } from "@/lib/trpc/server";
+import { HydrateClient } from "@/lib/trpc/server";
 
 import { Tabs } from "../tabs";
 import { Breadcrumb } from "./breadcrumb";
 import { NavActions } from "./nav-actions";
 
-export default async function Layout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const queryClient = getQueryClient();
-  await queryClient.prefetchQuery(trpc.workspace.get.queryOptions());
-
+export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <HydrateClient>
       <div>
