+++ apps/dashboard/src/app/(dashboard)/status-pages/[id]/edit/layout.tsx
@@ -1,0 +1,9 @@
+import { HydrateClient } from "@/lib/trpc/server";
+
+export default async function Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <HydrateClient>{children}</HydrateClient>;
+}
