+++ apps/dashboard/src/app/(dashboard)/status-pages/[id]/layout.tsx
@@ -4,12 +4,7 @@ import {
   AppHeaderContent,
 } from "@/components/nav/app-header";
 import { AppSidebarTrigger } from "@/components/nav/app-sidebar";
-import {
-  HydrateClient,
-  fetchQueryOrNotFound,
-  getQueryClient,
-  trpc,
-} from "@/lib/trpc/server";
+import { HydrateClient, fetchQueryOrNotFound, trpc } from "@/lib/trpc/server";
 
 import { Breadcrumb } from "./breadcrumb";
 import { NavActions } from "./nav-actions";
@@ -23,12 +18,10 @@ export default async function Layout({
   params: Promise<{ id: string }>;
 }) {
   const { id } = await params;
-  const queryClient = getQueryClient();
 
   await fetchQueryOrNotFound(
     trpc.page.get.queryOptions({ id: Number.parseInt(id) }),
   );
-  await queryClient.prefetchQuery(trpc.monitor.list.queryOptions());
 
   return (
     <HydrateClient>
